+++ crates/objc2/src/lib.rs
@@ -170,7 +170,7 @@
 #![warn(clippy::missing_errors_doc)]
 #![warn(clippy::missing_panics_doc)]
 // Update in Cargo.toml as well.
-#![doc(html_root_url = "https://docs.rs/objc2/0.4.1")]
+#![doc(html_root_url = "https://docs.rs/objc2/0.5.0")]
 
 #[cfg(not(feature = "alloc"))]
 compile_error!("The `alloc` feature currently must be enabled.");